pyproject.toml
[tool.poetry]
name = "playground"
version = "0.1.0"
description = ""
authors = ["Jared Sulzdorf <j.sulzdorf@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.0"
matplotlib = "^3.8.2"
gunicorn = "^21.2.0"
pyenchant = "^3.2.2"
pyspellchecker = "^0.7.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
